requirements.txt
fastmcp
neo4j
aiohttp
langchain
langchain-openai
langgraph
streamlit
plotly
pandas
python-dotenv